utils module for decoders
